generate embeddings enpoint using transformerjs
