+++ src/app/api/embed/route.js
@@ -1,8 +1,7 @@
-// Create a custom request handler for the /classify route.
-// For more information, see https://nextjs.org/docs/app/building-your-application/routing/router-handlers
-
 import { NextResponse } from 'next/server'
 import PipelineSingleton from './pipeline.js';
+
+
 
 export async function GET(request) {
     const text = request.nextUrl.searchParams.get('text');
@@ -13,9 +12,10 @@ export async function GET(request) {
     }
     // Get the classification pipeline. When called for the first time,
     // this will load the pipeline and cache it for future use.
-    const classifier = await PipelineSingleton.getInstance();
+    const embeddings = await PipelineSingleton.getInstance();
+
     // Actually perform the classification
-    const result = await classifier(text, {
+    const result = await embeddings(text, {
         pooling: 'mean',
         normalize: true,
     });
